requirements.txt
Pillow==11.3.0
boto3==1.39.11